requirements.txt
telethon==1.39.0
psycopg2-binary==2.8.2
python-dotenv==0.14.0